Create PortifolioObjects.js file
