+++ src/components/projects/PortfolioObjects.jsx
@@ -1,0 +1,73 @@
+const portfolioObjects = [
+  {
+    id: 1,
+    projectTitle: 'project1',
+    projectTitleDesktop: 'LeaderBoard',
+    tags: ['HTML', 'Webpack', 'Es6', 'API'],
+    description:
+      'The leaderboard website displays scores submitted by different players. It also allows you to submit your score. All data is preserved thanks to the external Leaderboard API service.',
+    liveUrl: 'https://sonickmumba.github.io/Leaderboard/',
+    sourceUrl: 'https://github.com/Sonickmumba/Leaderboard',
+    image: './images/projects/plash.jpg',
+  },
+
+  {
+    id: 2,
+    projectTitle: 'project2',
+    projectTitleDesktop: 'Meal App',
+    tags: ['Html', 'CSS', 'JavaScript'],
+    description:
+      'Web application based on an external API. We selected an API that provides data about Meals. One is able to like and comment on a meal.',
+    liveUrl: 'https://megha-n-bodke.github.io/CapstoneProject2/dist/',
+    sourceUrl: 'https://github.com/Sonickmumba/CapstoneProject2',
+    image: './images/projects/sa.jpg',
+  },
+
+  {
+    id: 3,
+    projectTitle: 'project3',
+    projectTitleDesktop: 'Awesome Books app',
+    tags: ['HTML', 'ES6', 'React'],
+    description:
+      'Awesome books is a simple website that displays a list of books and allows you to add and remove books from that list.',
+    liveUrl: 'https://sonickmumba.github.io/Awesome-Books-ES6/',
+    sourceUrl: 'https://github.com/Sonickmumba/Awesome-Books-ES6',
+    image: './images/projects/books.jpg',
+  },
+  {
+    id: 4,
+    projectTitle: 'project4',
+    projectTitleDesktop: 'Vespa Cars Rental',
+    tags: ['HTML', 'Ruby on Rails', 'React'],
+    description:
+      'Vesper application for all lovers of cars. Let your car make money for you by listing it on our platform so others can view and rent.',
+    liveUrl: 'https://vesper-4yj8.onrender.com/',
+    sourceUrl: 'https://github.com/Sonickmumba/vesper-frontend',
+    image: './images/projects/do.jpg',
+  },
+
+  {
+    id: 5,
+    projectTitle: 'project5',
+    projectTitleDesktop: 'Online Math Calculator',
+    tags: ['HTML', 'Es6', 'React'],
+    description:
+      'Single Page App for math calculations & reading random math related quotes, offering quick access to simple tools and resources',
+    liveUrl: 'https://dancing-conkies-c08f95.netlify.app/',
+    sourceUrl: 'https://github.com/Sonickmumba/math-magicians',
+    image: './images/projects/cal.ai',
+  },
+  {
+    id: 6,
+    projectTitle: 'project6',
+    projectTitleDesktop: 'ZAMSA-Seminar-Website',
+    tags: ['HTML', 'Es6', 'JavaScript'],
+    description:
+      'Zambia Mathematics Association Conference website with info on dates, speakers, tickets, location. Impact of Mathematics in a society.',
+    liveUrl: 'https://sonickmumba.github.io/Capstone-Project-1/',
+    sourceUrl: 'https://github.com/Sonickmumba/Capstone-Project-1',
+    image: './images/projects/conf2.png',
+  },
+];
+
+export default portfolioObjects;
